Moved duplicated code to one place
