+++ core/include/traccc/clusterization/measurement_creation.hpp
@@ -7,6 +7,7 @@
 
 #pragma once
 
+#include "traccc/clusterization/measurement_creation_helper.hpp"
 #include "traccc/definitions/primitives.hpp"
 #include "traccc/edm/cell.hpp"
 #include "traccc/edm/cluster.hpp"
@@ -82,23 +83,9 @@ struct measurement_creation
             }
 
             const auto &cl_id = clusters.at(0).header;
-            for (const auto &cell : cluster) {
-                scalar weight = signal_cell_modelling(cell.activation, cl_id);
-                if (weight > cl_id.threshold) {
-                    totalWeight += cell.activation;
-                    const point2 cell_position =
-                        position_from_cell(cell, cl_id);
+            for (const auto &cell : cluster)
+                calc_cluster_properties(cell, cl_id, mean, var, totalWeight);
 
-                    const point2 prev = mean;
-                    const point2 diff = cell_position - prev;
-
-                    mean = prev + (weight / totalWeight) * diff;
-                    for (std::size_t i = 0; i < 2; ++i) {
-                        var[i] = var[i] + weight * (diff[i]) *
-                                              (cell_position[i] - mean[i]);
-                    }
-                }
-            }
             if (totalWeight > 0.) {
                 measurement m;
                 // normalize the cell position
