+++ device/sycl/include/traccc/sycl/clusterization/measurement_creation.hpp
@@ -11,6 +11,7 @@
 #include "traccc/sycl/utils/queue_wrapper.hpp"
 
 // Project include(s).
+#include "traccc/clusterization/measurement_creation_helper.hpp"
 #include "traccc/definitions/primitives.hpp"
 #include "traccc/edm/cell.hpp"
 #include "traccc/edm/cluster.hpp"
