+++ device/sycl/src/clusterization/measurement_creation.sycl
@@ -106,23 +106,10 @@ void measurement_creation::operator()(
                     return;
                 }
 
-                for (const auto &cell : cluster) {
-                    scalar weight =
-                        signal_cell_modelling(cell.activation, cl_id);
-                    if (weight > cl_id.threshold) {
-                        totalWeight += cell.activation;
-                        const point2 cell_position =
-                            position_from_cell(cell, cl_id);
-                        const point2 prev = mean;
-                        const point2 diff = cell_position - prev;
+                for (const auto &cell : cluster)
+                    calc_cluster_properties(cell, cl_id, mean, var,
+                                            totalWeight);
 
-                        mean = prev + (weight / totalWeight) * diff;
-                        for (std::size_t i = 0; i < 2; ++i) {
-                            var[i] = var[i] + weight * (diff[i]) *
-                                                  (cell_position[i] - mean[i]);
-                        }
-                    }
-                }
                 if (totalWeight > 0.) {
                     measurement m;
                     // normalize the cell position
